Add WIFI init
This inits the power gpio. For compat-wireless (2.6.38-rc7) this
seemed not to be needed, but the (otherwise non working) staging
driver caused problems without it. So be sane and enable it.

Also add the gpio for the WIFI LED (to be enabled somewhere else)
and the RST (where do we need this?).