update defconfig again
the config is basicly based on the tegra chromium default config +
some drivers needed for ac100 minus some drivers not needed from
chromium, while the older config was based on android. In the end,
the kernel is larger and has lots of modules.
If you want a smaller kernel, remove some options.
If you find problems, send me a notice.