CHROMIUM: video: tegra: Add skeleton support for extensions
This adds the infrasturcture for an enhanced driver interface to program
extended capabilities of the Tegra display controller.

It exposes a new set of device nodes for userspace clients distinct from
the traditional fbdev device nodes.  This is necessary due to
limitations in the fbdev infrastructure that don't allow drivers to
store file-private data.

BUG=chrome-os-partner:1812
TEST=After the patch series, ran X server on Seaboard, verified HDMI hotplug works with xrandr
Signed-off-by: Robert Morell <rmorell@nvidia.com>
Signed-off-by: Olof Johansson <olofj@chromium.org>

Review URL: http://codereview.chromium.org/6676048

Change-Id: I3933f50bb2ad4b079c3d0480f1c661cafcbbfd46