nvec: rewrite isr and other small cleanups
The interrupt routine was rewritten. It is now based on the irq status
instead of some state variable. The tx queue is now driven by completions.
The nvec_send_* routines are protected by mutex now - maybe overkill,
but I wanted to be sure.

A possible memory corruption was fixed in async write, and timeouts
were added to the completion calls.

There is still some "ec confusion" happening, but it seems to recover.
All in all, it looks more stable for larger amount of machines/people.

As always: testers welcome!