Initialize lp0_vec start/size from atags
lp0_vec is a small code area provided the bootloader
This binary blob is used to suspend/resume the AVP chip.

Start and size of this area is normally passed to the kernel
through nvidia specific atags, so that's it's no longer mandatory
to pass the lp0_vec= parameter on the boot cmdline.