Disable power_down_on_bus_suspend for USB bus 3.
From the commit message of ehci-tegra on android.git.kernel.org:
"For additional power saving, the driver supports powering down the
phy on bus suspend when it is used, for example, to connect an internal
device that use an out-of-band remote wakeup mechanism (e.g. a gpio)."

Powering down the phy of bus 3 through this mechanism disables the
detection of newly connected devices to the external USB port,
therefore disabling its hotplug functionality.