CHROMIUM: video: tegra: fb: Remove yres_virtual hack
This hack was added for Android to be able to perform panning.  However,
not all clients want this, and for those that don't, it generally
doubles the amount of memory required.

Instead, the client specifying the mode should be the one doubling
yres_virtual if it is desired.

Signed-off-by: Robert Morell <rmorell@nvidia.com>

BUG=chrome-os-partner:4249
TEST=Verify that fbconsole VTs still work, X11 still works, and VT
switching between the two still works

Change-Id: I723abb07d92a7588566f73266678176a7d1a01ab
Reviewed-on: http://gerrit.chromium.org/gerrit/4862
Tested-by: Robert Morell <rmorell@nvidia.com>
Reviewed-by: Doug Anderson <dianders@chromium.org>