staging: nvec: rewrite of the ISR and data handling
This is a rather large patch. The older code wasn't stable enough
in all cases as it confused the EC during high system load. The new
code uses a better state machine which is quiet similar to the
original code from NVIDIA.

In short, this introduces:
	- new ISR state machine for protocol handling
	- a ring buffer for received messages
	  (this avoids the use of malloc in the ISR)
	- the list operations are now protected by spinlocks to
	  handle concurrent access in the bottom half
	- a "nvec" workqueue. The EC is very sensitive regarding
	  correct timing. In the past, system events often caused
	  slow responses resulting in a protocol timeout. The new code
	  avoids this by using its own workqueue.

Things to do:

* the ring buffer code is not finished yet, e.g. the received link
list is just a pointer to the ring buffer array, while it should
be a list of its own.

* error handling is also not finished yet, e.g. during ring buffer
overrun the ec needs to be reseted.

* the locking needs a review.

* the sync_write should return an error code instead of the received
message.

* other things I've forgotten now

Signed-off-by: Marc Dietrich <marvin24@gmx.de>