nvec: try to fix new init
We initilized two times, first in i2c-tegra and later in nvec. Even
worse, i2c-tegra got it wrong. So, fix i2c-tegra and remove the init
stuff from nvec.

I still got some timeouts, but guess what - .32 also timeouts during
boot (but is able to recover). So, time to concentrate on how to
recover probably...

Upstreams says that we need to use more of i2c-tegra code, so we may
use i2c-tegra's versions of readl and writel to get rid of the i2c_dev
struct share. Volunteers?