CHROMIUM: video: tegra: Honor swap_interval field
The swap_interval field should be honored now.  0 means to perform the flip
as soon as possible (we use the H_BLANK interrupt to trigger this).  Greater
than 0 means to wait that many frames before performing the flip (we use
the FRAME_END interrupt to trigger this).

BUG=chrome-os-partner:3732, chrome-os-partner:5903
TEST=enabled ARGB cursors in my ChromeOS filesystem.  Observed the laggy
mouse behavior.  Observed that the laggy mouse behavior is gone after this
patch.  Hacked swap_interval for main window (window A) with values of 0
through 5, and made sure there weren't any visibly noticeable regressions.

Change-Id: I3f0f47a28a2af476fef91d1a63e28b5dc8f593dd
Signed-off-by: Andrew Chew <achew@nvidia.com>
Reviewed-on: http://gerrit.chromium.org/gerrit/7818
Reviewed-by: Olof Johansson <olofj@chromium.org>
Tested-by: Olof Johansson <olofj@chromium.org>