nvec: use init from i2c-tegra
This patches nvec to make use of the slave controller initialization in
i2c-tegra. For this to work, a small patch has to be applied to the latter.

The movement of the whole i2c tegra device struct to the header file is not
so nice. A possible (mid term) solution would be to split the device struct
in some resource part (used by both drivers) and a protocol part which is
kept inside the drivers itself.

A more long term solution could be to make nvec not to deal with the slave
controller resources at all. This definitely needs more discussion.