partitions: add nvtegra support
Add support for Tegra/Android like partition tables. In newer kernels
(>= 3.0) access to the "hidden" part of a mmc is permitted via one or
more block devices mmcblk(N)boot(M), with N the device number (0 for
internal mmc) and M the number of the hidden part (0,1,2,...).

TODO: currently the boot partitions containing the partitons table
are scanned as usual, but the data must be preserved and used by the
actual device.