module_param: make bool parameters really bool (net & drivers/net)
module_param(bool) used to counter-intuitively take an int.  In
fddd5201 (mid-2009) we allowed bool or int/unsigned int using a messy
trick.

It's time to remove the int/unsigned int option.  For this version
it'll simply give a warning, but it'll break next kernel version.

(Thanks to Joe Perches for suggesting coccinelle for 0/1 -> true/false).

Cc: "David S. Miller" <davem@davemloft.net>
Cc: netdev@vger.kernel.org
Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>
Signed-off-by: David S. Miller <davem@davemloft.net>

Conflicts:

	drivers/net/ethernet/amd/amd8111e.h
	drivers/net/ethernet/chelsio/cxgb4/cxgb4_main.c
	drivers/net/ethernet/mellanox/mlx4/fw.c
	drivers/net/ethernet/mellanox/mlx4/main.c
	drivers/net/ethernet/via/via-rhine.c
	drivers/net/wireless/ath/ath5k/ath5k.h
	net/bluetooth/hci_event.c
	net/dccp/ccids/ccid2.c

Signed-off-by: Marc Dietrich <marvin24@gmx.de>