CHERRY-PICK: drm/i915: Only look for matching clocks for LVDS downclock
This patch enforces that the downclock clock source is the same as the preferred
clock source for LVDS. This fixes a bug where the driver chooses a downclock
clock source with a different P than the preferred mode clock source. This
happened even if the preferred clock source implemented an acceptable rate for
the downclock. The result of this bug is that downclock is disabled.

Signed-off-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Jesse Barnes <jbarnes@virtuousgeek.org>
Signed-off-by: Daniel Vetter <daniel.vetter@ffwll.ch>
(cherry picked from commit cec2f356d59d9e070413e5966a3c5a1af136d948)

BUG=chromium-os:9803
TEST=Tested on alex and lumpy
Change-Id: I7f709dee1e64f004db3cbd7516b90e450e39e19b
Reviewed-on: https://gerrit.chromium.org/gerrit/14299
Reviewed-by: Olof Johansson <olofj@chromium.org>
Commit-Ready: Sean Paul <seanpaul@chromium.org>
Tested-by: Sean Paul <seanpaul@chromium.org>