ARM: tegra: paz00: add a missing clock for disp1
From: Paul Fertser <fercerpav@gmail.com>

Without this the disp1 clock is wrongly parented (if wasn't initialised
by a bootloader) and the panel starts working in 14Hz V refresh mode.
Adding this allows to use the proper 52Hz mode.

This may not be required if u-boot sets up the clock tree already, but that
is debatable (with upstream).

Signed-off-by: Paul Fertser <fercerpav@gmail.com>
Signed-off-by: Marc Dietrich <marvin24@gmx.de>