staging: nvec: fix bus clock
This was set to a too high value in the clock table. Also later on
the nvec didn't got the right clock at all (used the controller clock)
instead of the bus clock. Urg.

I failed to used clk_add_alias so I just added a dup clock as the
vibrante slave driver does it for tegra3.

Signed-off-by: Marc Dietrich <marvin24@gmx.de>