staging: nvec: re-enable the clock on resume
On resume the slave controller is reinitialized. The tegra i2c master
controller disables the clock at the end of the initialiation, propably
to save some power, and enables it again on each transfer. We don't
do this yet and also forgot to enable the clock on resume. Fix this
copy-paste error by not disabling the clock after initialization.

Signed-off-by: Marc Dietrich <marvin24@gmx.de>